Refactor: `EventsArchive`
Static classes to instances.
